refactor:migrate routes in controllers

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Models\Quote;
 use App\Http\Controllers\MovieController;
 use App\Http\Controllers\QuoteController;
 use App\Http\Controllers\LoginController;
+use App\Http\Controllers\MainController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,35 +19,20 @@ use App\Http\Controllers\LoginController;
 |
 */
 
-Route::get('/', function () {
-	// return view('main',['movies'=>Movie::inRandomOrder()->get()]);
-	// $quantity = Movie::count();
-	if(Movie::all()->count()){
-   
-		return view('main', ['movies'=>Quote::all()->random()]);
-	}else{
-		return 'Database is empty :)';
-	}
-});
-Route::get('quotes/{slug}', function ($slug) {
-	// dd($slug);
-	return view('movies', ['movies'=>Movie::all(), 'slug'=>$slug]);
-});
+Route::get('/', [MainController::class,'show']);
+Route::get('quotes/{slug}', [MainController::class,'showQuotes']);
 
 
-Route::get('/choose',function(){
-	return view('choose');
-});
+// Route::get('/choose',function(){
+// 	return view('choose');
+// });
 
 Route::get('add/movie',[MovieController::class,'show']);
 Route::post('add/movie',[MovieController::class,'store']);
 
 
 Route::post('add/quote',[QuoteController::class,'store']);
-Route::get('add/quote',function(){
-
- return view('addQuote',['quotes'=>Quote::all(),'allMovie'=>Movie::all()]);
-});
+Route::get('add/quote',[QuoteController::class, 'show']);
 
 
 Route::get('login',[LoginController::class,'create'])->middleware('guest');
@@ -55,12 +41,12 @@ Route::post('login',[LoginController::class,'store'])->middleware('guest');
 
 Route::post('logout',[LoginController::class,'destroy'])->middleware('auth');
 
-Route::delete('admin/quotes/{quote}',[QuoteController::class,'destroy']);
-Route::delete('admin/movies/{movie}',[MovieController::class,'destroy']);
+Route::delete('admin/quotes/{quote}',[QuoteController::class,'destroy'])->middleware('auth');
+Route::delete('admin/movies/{movie}',[MovieController::class,'destroy'])->middleware('auth');
 
 
-Route::get('admin/quotes/{quote}/edit',[QuoteController::class,'edit']);
-Route::patch('admin/quotes/{quote}',[QuoteController::class,'update']);
+Route::get('admin/quotes/{quote}/edit',[QuoteController::class,'edit'])->middleware('auth');
+Route::patch('admin/quotes/{quote}',[QuoteController::class,'update'])->middleware('auth');
 
-Route::get('admin/movies/{movie}/edit',[MovieController::class,'edit']);
-Route::patch('admin/movies/{movie}',[MovieController::class,'update']);
+Route::get('admin/movies/{movie}/edit',[MovieController::class,'edit'])->middleware('auth');
+Route::patch('admin/movies/{movie}',[MovieController::class,'update'])->middleware('auth');
